Add files via upload
This project showcases my work in healthcare analytics and machine learning, with a particular emphasis on dementia prediction and data-driven insights. This project focuses on leveraging data analytics and machine learning techniques to predict dementia onset. By analyzing patient data and identifying key indicators, the project aims to develop a predictive model that can forecast the likelihood of dementia in individuals. The project includes data preprocessing, exploratory analysis, feature engineering, model building, and evaluation steps. Through this work, we aim to contribute to the field of healthcare analytics by developing a tool that can assist healthcare professionals in early detection and intervention for dementia.